fix(routing): move console to dashboard
Serve the previous dashboard at /legacy-dashboard and redirect legacy /console URLs to the primary UI.
