+++ central/hooks.py
@@ -11,13 +11,11 @@ fixtures = [
 	{"dt": "Role", "filters": [["name", "in", ["Central User"]]]},
 ]
 
-# Central Console SPAs — each client route resolves to its www entry page.
-#   /dashboard → legacy JS app (www/dashboard).
-#   /console   → new TypeScript app (www/console). Both coexist; routing is the
-#               only thing that keeps them apart.
+# The TypeScript UI owns the product route. Keep the previous dashboard
+# available while its remaining surfaces are migrated.
 website_route_rules = [
 	{"from_route": "/dashboard/<path:app_path>", "to_route": "dashboard"},
-	{"from_route": "/console/<path:app_path>", "to_route": "console"},
+	{"from_route": "/legacy-dashboard/<path:app_path>", "to_route": "legacy-dashboard"},
 ]
 
 # Apps
