--- central/www/console.py
@@ -1,30 +1,0 @@
-from __future__ import annotations
-
-import frappe
-from frappe.sessions import get_csrf_token
-
-no_cache = 1
-
-
-def get_context(context):
-	"""Boot data for the new Central Console SPA (TypeScript / frappe-ui beta).
-
-	The page itself is generated by the frappe-ui Vite build (`yarn build` copies
-	the hashed `index.html` to this `www/console.html`). The build's jinjaBootData
-	plugin injects a loop that sets `window[key] = boot[key]` for each key below, so
-	the SPA can read `window.csrf_token`, `window.user`, etc. at load time.
-	frappe-ui's request layer sends `window.csrf_token` as the X-Frappe-CSRF-Token
-	header, so its POST (write) API calls aren't rejected.
-
-	Mirrors www/dashboard.py — the two SPAs share the same boot contract so either
-	can be served independently while we migrate.
-	"""
-	context.no_cache = 1
-	# init_context() already set context.boot via get_website_settings(); extend it
-	# rather than replace so framework defaults survive alongside the SPA's auth
-	# essentials.
-	boot = context.boot or frappe._dict()
-	boot["csrf_token"] = get_csrf_token()
-	boot["user"] = frappe.session.user
-	context.boot = boot
-	return context
