+++ central/www/dashboard.html
@@ -4,14 +4,8 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Central Console</title>
-    <!-- Stripe.js must be served from js.stripe.com (PCI; it can't be bundled).
-         Loading it here (in parallel with the app, not on dialog-open) means it's
-         already cached when the Add-card dialog needs it — so a slow connection
-         doesn't leave the card widget blank. @stripe/stripe-js reuses this tag. -->
-    <link rel="preconnect" href="https://js.stripe.com" crossorigin />
-    <script src="https://js.stripe.com/v3/" async></script>
-    <script type="module" crossorigin src="/assets/central/dashboard/assets/index-BDkRDmw1.js"></script>
-    <link rel="stylesheet" crossorigin href="/assets/central/dashboard/assets/index-BulsytHD.css">
+    <script type="module" crossorigin src="/assets/central/dashboard/assets/index-tvPi813H.js"></script>
+    <link rel="stylesheet" crossorigin href="/assets/central/dashboard/assets/index-C8QcWoSd.css">
   </head>
   <body>
     <div id="app"></div>
