+++ central/www/dashboard.py
@@ -7,9 +7,9 @@ no_cache = 1
 
 
 def get_context(context):
-	"""Boot data for the Central console SPA.
+	"""Boot data for the primary Central dashboard SPA.
 
-	The page itself is generated by the frappe-ui Vite build (`yarn build`
+	The page is generated by the TypeScript frappe-ui build (`yarn build`
 	copies the hashed `index.html` to this `www/dashboard.html`). The build's
 	jinjaBootData plugin injects a loop that sets `window[key] = boot[key]` for
 	each key below, so the SPA can read `window.csrf_token`, `window.user`, etc.
