+++ console/src/main.ts
@@ -5,7 +5,7 @@ import App from '@/App.vue'
 import './style.css'
 
 // The new data-fetching composables (useCall/useList) read window.csrf_token —
-// injected by central/www/console.py — and POST relative to the served origin,
+// injected by central/www/dashboard.py — and POST relative to the served origin,
 // so no extra request config is needed beyond the FrappeUI plugin.
 const app = createApp(App)
 app.use(router)
