+++ package.json
@@ -8,14 +8,17 @@
   "type": "module",
   "scripts": {
     "postinstall": "cd dashboard && yarn install && cd ../console && yarn install",
-    "dev": "cd dashboard && yarn dev",
+    "dev": "cd console && yarn dev",
     "dev:console": "cd console && yarn dev",
+    "dev:legacy-dashboard": "cd dashboard && yarn dev",
     "test": "cd dashboard && yarn test",
     "watch": "yarn dev",
     "build": "NODE_ENV=production yarn run build-all",
-    "build-all": "yarn run build-app && yarn run build-console",
-    "build-app": "cd dashboard && yarn build",
-    "build-console": "cd console && yarn build",
+    "build-all": "yarn run build-legacy-dashboard && yarn run build-dashboard",
+    "build-dashboard": "cd console && yarn build",
+    "build-legacy-dashboard": "cd dashboard && yarn build",
+    "build-app": "yarn run build-legacy-dashboard",
+    "build-console": "yarn run build-dashboard",
     "test:e2e": "playwright test",
     "test:e2e:headed": "playwright test --headed",
     "test:e2e:report": "playwright show-report",
